Issue: https://github.com/socialsensingbot/frontend/issues/105 Issue Description: Limit on number of tweets which can be ignored?
First draft of using tabs.
